backend/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
et-xmlfile==1.1.0
fonttools==4.51.0
gunicorn==22.0.0
h11==0.14.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
waitress==3.0.0
whitenoise==6.6.0